requirements.txt
ipykernel
pandas
plotly.express
seaborn
scikit-learn
numpy
streamlit
xgboost
catboost
dill
PyYAML
hydra-core
# -e .